Flatten data storage types: ConfigStore, StateStore, SchemaVersionStore (#1289)
The rest of the scheduler depends heavily on current behavior in these, it's infeasible to have alternate implementations at this level. In practice other implementations would be at the key/value Persister level instead.
