+++ frameworks/cassandra/src/test/java/com/mesosphere/sdk/cassandra/scheduler/CassandraRecoveryPlanOverriderTest.java
@@ -1,6 +1,5 @@
 package com.mesosphere.sdk.cassandra.scheduler;
 
-import com.mesosphere.sdk.config.ConfigStore;
 import com.mesosphere.sdk.scheduler.plan.DefaultPodInstance;
 import com.mesosphere.sdk.scheduler.plan.Phase;
 import com.mesosphere.sdk.scheduler.plan.Plan;
@@ -8,8 +7,7 @@ import com.mesosphere.sdk.scheduler.plan.PodInstanceRequirement;
 import com.mesosphere.sdk.scheduler.recovery.RecoveryType;
 import com.mesosphere.sdk.specification.*;
 import com.mesosphere.sdk.specification.yaml.RawServiceSpec;
-import com.mesosphere.sdk.state.DefaultConfigStore;
-import com.mesosphere.sdk.state.DefaultStateStore;
+import com.mesosphere.sdk.state.ConfigStore;
 import com.mesosphere.sdk.state.StateStore;
 import com.mesosphere.sdk.state.StateStoreUtils;
 import com.mesosphere.sdk.storage.MemPersister;
@@ -64,8 +62,8 @@ public class CassandraRecoveryPlanOverriderTest extends BaseServiceSpecTest {
     @Before
     public void beforeEach() throws Exception {
         super.beforeEach();
-        stateStore = new DefaultStateStore(new MemPersister());
-        ConfigStore<ServiceSpec> configStore = new DefaultConfigStore<>(
+        stateStore = new StateStore(new MemPersister());
+        ConfigStore<ServiceSpec> configStore = new ConfigStore<>(
                 DefaultServiceSpec.getConfigurationFactory(getServiceSpec()),
                 new MemPersister());
         UUID targetConfig = configStore.store(getServiceSpec());
