+++ sdk/scheduler/src/main/java/com/mesosphere/sdk/state/SchedulerState.java
@@ -53,4 +53,3 @@ public class SchedulerState extends DefaultObservable {
         return stateStore;
     }
 }
-
