+++ sdk/scheduler/src/main/java/com/mesosphere/sdk/state/StateStoreUtils.java
@@ -1,7 +1,6 @@
 package com.mesosphere.sdk.state;
 
 import com.google.protobuf.TextFormat;
-import com.mesosphere.sdk.config.ConfigStore;
 import com.mesosphere.sdk.config.ConfigurationUpdater;
 import com.mesosphere.sdk.offer.MesosResource;
 import com.mesosphere.sdk.offer.TaskException;
@@ -11,7 +10,6 @@ import com.mesosphere.sdk.specification.PodInstance;
 import com.mesosphere.sdk.specification.ServiceSpec;
 import com.mesosphere.sdk.specification.TaskSpec;
 import com.mesosphere.sdk.storage.StorageError.Reason;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.mesos.Protos;
 import org.apache.mesos.Protos.TaskInfo;
 import org.apache.mesos.Protos.TaskStatus;
@@ -33,15 +31,10 @@ public class StateStoreUtils {
     private static final String UNINSTALLING_PROPERTY_KEY = "uninstalling";
     private static final String LAST_COMPLETED_UPDATE_TYPE_KEY = "last-completed-update-type";
     private static final String PROPERTY_TASK_INFO_SUFFIX = ":task-status";
-    private static final int MAX_VALUE_LENGTH_BYTES = 1024 * 1024; // 1MB
 
     private StateStoreUtils() {
         // do not instantiate
     }
-
-
-    // Utilities for StateStore users:
-
 
     /**
      * Returns the requested property, or an empty array if the property is not present.
@@ -53,7 +46,6 @@ public class StateStoreUtils {
             return new byte[0];
         }
     }
-
 
     /**
      * Fetches and returns all {@link TaskInfo}s for tasks needing recovery.
@@ -117,26 +109,6 @@ public class StateStoreUtils {
         return results;
     }
 
-
-    // Utilities for StateStore implementations:
-
-
-    /**
-     * Shared implementation for validating property key limits, for use by all StateStore
-     * implementations.
-     *
-     * @see StateStore#storeProperty(String, byte[])
-     * @see StateStore#fetchProperty(String)
-     */
-    public static void validateKey(String key) throws StateStoreException {
-        if (StringUtils.isBlank(key)) {
-            throw new StateStoreException(Reason.LOGIC_ERROR, "Key cannot be blank or null");
-        }
-        if (key.contains("/")) {
-            throw new StateStoreException(Reason.LOGIC_ERROR, "Key cannot contain '/'");
-        }
-    }
-
     /**
      * Verifies that the supplied TaskStatus corresponds to a single TaskInfo in the provided StateStore and returns the
      * TaskInfo.
@@ -168,23 +140,6 @@ public class StateStoreUtils {
         }
 
         return taskInfoOptional.get();
-    }
-
-    /**
-     * Shared implementation for validating property value limits, for use by all StateStore
-     * implementations.
-     *
-     * @see StateStore#storeProperty(String, byte[])
-     */
-    public static void validateValue(byte[] value) throws StateStoreException {
-        if (value == null) {
-            throw new StateStoreException(Reason.LOGIC_ERROR, "Property value must not be null.");
-        }
-        if (value.length > MAX_VALUE_LENGTH_BYTES) {
-            throw new StateStoreException(Reason.LOGIC_ERROR, String.format(
-                    "Property value length %d exceeds limit of %d bytes.",
-                    value.length, MAX_VALUE_LENGTH_BYTES));
-        }
     }
 
     public static Collection<Protos.Resource> getReservedResources(Collection<Protos.Resource> resources) {
