+++ sdk/scheduler/src/test/java/com/mesosphere/sdk/api/ConfigResourceTest.java
@@ -1,8 +1,8 @@
 package com.mesosphere.sdk.api;
 
-import com.mesosphere.sdk.config.ConfigStore;
-import com.mesosphere.sdk.config.ConfigStoreException;
 import com.mesosphere.sdk.config.StringConfiguration;
+import com.mesosphere.sdk.state.ConfigStore;
+import com.mesosphere.sdk.state.ConfigStoreException;
 import com.mesosphere.sdk.storage.StorageError.Reason;
 import org.json.JSONArray;
 import org.junit.Before;
