+++ sdk/scheduler/src/test/java/com/mesosphere/sdk/scheduler/uninstall/UninstallSchedulerDeregisterTest.java
@@ -1,8 +1,8 @@
 package com.mesosphere.sdk.scheduler.uninstall;
 
-import com.mesosphere.sdk.config.ConfigStore;
 import com.mesosphere.sdk.specification.ServiceSpec;
-import com.mesosphere.sdk.state.DefaultStateStore;
+import com.mesosphere.sdk.state.ConfigStore;
+import com.mesosphere.sdk.state.StateStore;
 import com.mesosphere.sdk.storage.MemPersister;
 
 import org.junit.Before;
@@ -21,7 +21,7 @@ public class UninstallSchedulerDeregisterTest {
     public void beforeEach() throws Exception {
         // No framework ID is set yet, and there are no tasks, and no SchedulerDriver
         uninstallScheduler = new UninstallScheduler(
-                0, Duration.ofSeconds(1), new DefaultStateStore(new MemPersister()), configStore);
+                0, Duration.ofSeconds(1), new StateStore(new MemPersister()), configStore);
     }
 
     @Test
